buy/sell loop

+++ alpaca_trade_api/models/aux.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import typing
 import uuid
-from decimal import Decimal
 from enum import Enum
 
 from pydantic import BaseModel, Field
@@ -31,5 +30,5 @@ class OrderBase(BaseModel):
     client_order_id: str = Field(default_factory=lambda: str(uuid.uuid4()))
     qty: int
     time_in_force: OrderTimeInFore = OrderTimeInFore.DAY
-    limit_price: typing.Optional[Decimal]
-    stop_price: typing.Optional[Decimal] = None
+    limit_price: typing.Optional[float]
+    stop_price: typing.Optional[float] = None
